refactor before adding unit test
