+++ package.json
@@ -36,6 +36,7 @@
     "graphql-tag": "^2.12.6",
     "js-waku": "^0.24.0",
     "node-fetch": "3.2.10",
+    "p-map": "^5.5.0",
     "protobufjs": "^7.0.0"
   },
   "lint-staged": {
